+++ consumer/src/main/java/com/projecAsapcard/consumer/model/Person.java
@@ -45,11 +45,11 @@ public class Person {
 		this.name = name;
 	}
 
-	public Integer getAge() {
+	public Short getAge() {
 		return age;
 	}
 
-	public void setAge(Integer age) {
+	public void setAge(Short age) {
 		this.age = age;
 	}
